设计游戏的存储结构
