+++ Plugins/SimpleArchives/Source/SimpleArchives/Public/UI_ArchivesSystem.h
@@ -55,6 +55,8 @@ protected:
 
 	//void ResetArchivesBar(UUI_ArchivesBar* InArchivesBar, const FSaveSlot *InData);
 
+	void Save(bool bCover);
+
 private:
 	void CallAllArchivesBar(TFunction<void(UUI_ArchivesBar* InArchivesBar)>);
 	void CallAllArchivesBarBreak(TFunction<bool(UUI_ArchivesBar* InArchivesBar)>);
